requirements.txt
adal==1.2.1
asn1crypto==0.24.0
astroid==2.1.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.5
django-cors-headers==2.5.0
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.0
docutils==0.14
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
idna==2.8
isodate==0.6.0
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.4
msrestazure==0.6.0
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==5.4.1
pipenv==2018.11.26
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.4.0
wrapt==1.11.1
